server_api/requirements.txt
# server_api dependencies
fastapi
pillow
torch
diffusers